move TestInMemoryWriterDataRace into index_test.go
